Created PHashInternal
